refactor(DTFS2-6113): improve naming of DTFS Central facilities API tests
